Fixes some proposals that wouldn't open.

+++ lib/screens/home/explorer/proposals_flow/proposals_details/deposit_bar_chart.dart
@@ -11,8 +11,8 @@ class DepositBarChart extends StatelessWidget {
     ),
   }) : super(key: key);
 
-  final int total;
-  final int current;
+  final double total;
+  final double current;
   final EdgeInsets padding;
 
   @override
